PlayGame shell added, Player Object added
PlayGame shell to be used for unit testing FOR NOW, Game now will
generate Chessboard, Pieces and Player Objects at the start of the game.
Player will contain the AI's or a request for a move.
Will not push anything more until I've figured out how to do unit
testing.